+++ packages/orq-rc/RELEASES.md
@@ -7539,3 +7539,13 @@ Based on:
 - [typescript v4.5.0-rc.41] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.41 - packages/orq-rc
+
+## 2026-03-09 14:40:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.1 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.42] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.42 - packages/orq-rc